refactor: refactor code & add comments
