+++ lib/features/ performance_survey/photo_gallery/photo_gallery_screen_listview.dart
@@ -1,6 +1,9 @@
 import 'package:faker/faker.dart';
 import 'package:flutter/material.dart';
+import 'package:tech_publish_app/features/%20performance_survey/photo_gallery/photo.dart';
 
+/// ListViewで[Photo]の一覧を表示させた画像ギャラリー画面
+///　ListViewは、リストの要素が固定されている場合に便利です
 class ListViewScreen extends StatelessWidget {
   const ListViewScreen({super.key});
 
@@ -16,7 +19,8 @@ class ListViewScreen extends StatelessWidget {
         // cacheExtent: 250,
         children: [
           for (var i = 0; i < _imageCount; i++)
-            _Photo(
+            Photo(
+              height: _photoHeight,
               url: faker.image.image(
                 keywords: ['nature', i.toString()],
                 random: true,
@@ -28,20 +32,5 @@ class ListViewScreen extends StatelessWidget {
   }
 }
 
-class _Photo extends StatelessWidget {
-  const _Photo({required this.url});
-
-  final String url;
-
-  @override
-  Widget build(BuildContext context) {
-    return Image.network(
-      url,
-      width: double.infinity,
-      height: 200,
-      fit: BoxFit.cover,
-    );
-  }
-}
-
 const _imageCount = 100;
+const double _photoHeight = 200;
